Fixed check previous monitoring session Alert cancel action
Previously when the X button in dialog was pressed, a new monitoring
session was created instead of being a no-op. This was because the X
button was being picked up as the same button action as the No button to
create a new session. Now a cancel button is created which fixes this
error.