Fixed modified file detection and added statistics to UI
Turns out comparing 2 FileTime instances is a bad idea as 1 instance
can contain truncated time which makes it impossible to check for
equality. Instead the FileTime is now converted to milliseconds which is
the correct way to do it and now no files are being detected as having
changed when they are identical.

A statistics UI view was added to provide visibility on how many files
were scanned etc.